Add files via upload
In this script, you'll need to replace your_email@example.com and your_email_password with your own email credentials. You can also customize the message and the list of recipients to fit your needs. When you run the script, it will send an email to each recipient in the list with the message "Wishing you a happy day! :)" and the subject "Sending you some happiness!".